+++ src/data/store.ts
@@ -70,6 +70,7 @@ export function data(
         variables: queryStoreValue.variables,
         store: clonedState,
         dataIdFromObject: config.dataIdFromObject,
+        fragmentMap: queryStoreValue.fragmentMap,
       });
 
       return newState;