+++ src/queries/getFromAST.ts
@@ -114,3 +114,18 @@ export function replaceOperationDefinition(doc: Document,
 
   return docCopy;
 }
+
+export interface FragmentMap {
+  [fragmentName: string]: FragmentDefinition;
+}
+
+// Utility function that takes a list of fragment definitions and makes a hash out of them
+// that maps the name of the fragment to the fragment definition.
+export function createFragmentMap(fragments: FragmentDefinition[]): FragmentMap {
+  const symTable: FragmentMap = {};
+  fragments.forEach((fragment) => {
+    symTable[fragment.name.value] = fragment;
+  });
+
+  return symTable;
+}