+++ test/client.ts
@@ -313,6 +313,7 @@ describe('client', () => {
             networkError: null,
             graphQLErrors: null,
             forceFetch: false,
+            fragmentMap: {},
             returnPartialData: false,
             lastRequestId: 1,
           },
@@ -569,6 +570,106 @@ describe('client', () => {
       done();
     });
 
+  });
+
+  it('should be able to handle named fragments on forced fetches', (done) => {
+    const query = gql`
+      fragment authorDetails on Author {
+        firstName
+        lastName
+      }
+      query {
+        author {
+          ...authorDetails
+        }
+      }`;
+    const result = {
+      'author': {
+        'firstName': 'John',
+        'lastName': 'Smith',
+      },
+    };
+    const networkInterface = mockNetworkInterface(
+    {
+      request: { query },
+      result: { data: result },
+    });
+    const client = new ApolloClient({
+      networkInterface,
+    });
+    client.query({ forceFetch: true, query }).then((actualResult) => {
+      assert.deepEqual(actualResult.data, result);
+      done();
+    });
+  });
+
+  it('should be able to handle named fragments with multiple fragments', (done) => {
+    const query = gql`
+      query {
+        author {
+          ...authorDetails
+          ...moreDetails
+        }
+      }
+      fragment authorDetails on Author {
+        firstName
+        lastName
+      }
+      fragment moreDetails on Author {
+        address
+      }`;
+    const result = {
+      'author' : {
+        'firstName': 'John',
+        'lastName': 'Smith',
+        'address': '1337 10th St.',
+      },
+    };
+
+    const networkInterface = mockNetworkInterface(
+    {
+      request: { query },
+      result: { data: result },
+    });
+    const client = new ApolloClient({
+      networkInterface,
+    });
+    client.query({ query }).then((actualResult) => {
+      assert.deepEqual(actualResult.data, result);
+      done();
+    });
+  });
+
+  it('should be able to handle named fragments', (done) => {
+    const query = gql`
+      query {
+        author {
+          ...authorDetails
+        }
+      }
+      fragment authorDetails on Author {
+        firstName
+        lastName
+      }`;
+    const result = {
+      'author' : {
+        'firstName': 'John',
+        'lastName': 'Smith',
+      },
+    };
+
+    const networkInterface = mockNetworkInterface(
+    {
+      request: { query },
+      result: { data: result },
+    });
+    const client = new ApolloClient({
+      networkInterface,
+    });
+    client.query({ query }).then((actualResult) => {
+      assert.deepEqual(actualResult.data, result);
+      done();
+    });
   });
 
   describe('accepts dataIdFromObject option', () => {